Build in a pure Wayland environment
Note that this change only allow the package to build in a pure wayland
environment but doesn't make it functionnal since some X11 -> Wayland
equivalents doesn't exist yet. The changes that remain to be made in
order to get it working in a pure Wayland environment are marked by the
comment "TO_DO_WAYLAND". Bug report regarding this issue: PTREL-770.

Change-Id: I8378bc856c0932f40e2ea7c67e9ddc653339fdfa
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>